requirements.txt
fastapi
numpy
pandas
pillow
uvicorn
python-multipart